+++ Editor/ValidationSuite/ValidationTests/ApiValidation.cs
@@ -229,8 +229,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests
                     };
 
                     const string logsDirectory = "Logs";
-                    if (!Directory.Exists(logsDirectory))
-                        Directory.CreateDirectory(logsDirectory);
+                    Utilities.EnsureDirectoryExists(logsDirectory);
 
                     File.WriteAllText($"{logsDirectory}/ApiValidationParameters.txt", $"previous: {oldAssemblyPath}\ncurrent: {info.assembly.outputPath}\nsearch path: {string.Join("\n", assemblySearchFolder)}");
                     var apiChangesAssemblyInfo = new APIChangesCollector.AssemblyInfo()