+++ Editor/ValidationSuite/ValidationTests/Standards/US0041-APIDocumentationIncluded.cs
@@ -3,15 +3,41 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using UnityEditor.Compilation;
 
 namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests.Standards
 {
-    class APIDocumentationIncludedUS0041 : BaseStandardChecker
+    class APIDocumentationIncludedUS0041 : BaseStandardChecker, IPvpChecker
     {
         public override string StandardCode => "US-0041";
         public override StandardVersion Version => new StandardVersion(2, 1, 1);
 
-        public void Check(string packagePath, AssemblyInfo[] assemblyInfo, ValidationAssemblyInformation validationAssemblyInformation)
+        public void Check(string packagePath, AssemblyInfo[] assemblyInfo, ValidationAssemblyInformation vai)
+        {
+            var filterYamlParameter = "";
+            var filterYamlPath = Path.Combine(packagePath, "Documentation~", "filter.yml");
+            if (Utilities.FileExists(filterYamlPath))
+            {
+                filterYamlParameter = $@"--path-to-filter-yaml=""{filterYamlPath}""";
+            }
+
+            try
+            {
+                var stdoutLines = CheckV1Impl(packagePath, filterYamlParameter, assemblyInfo, vai);
+                if (stdoutLines.Length > 0)
+                {
+                    var errorMessage = FormatErrorMessage(stdoutLines);
+                    AddWarning(errorMessage);
+                }
+            }
+            catch (InternalTestErrorException e)
+            {
+                AddError(e.Message);
+            }
+        }
+
+        // Subject to PVP stability guarantee.
+        static string[] CheckV1Impl(string packagePath, string filterYamlParameter, AssemblyInfo[] assemblyInfo, ValidationAssemblyInformation vai)
         {
             var monopath = Utilities.GetMonoPath();
             var exePath = Path.GetFullPath("packages/com.unity.package-validation-suite/Bin~/FindMissingDocs/FindMissingDocs.exe");
@@ -23,7 +49,7 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests.Standards
             foreach (var assembly in assemblyInfo)
             {
                 //exclude sources from test assemblies explicitly. Do not exclude entire directories, as there may be nested public asmdefs
-                if (validationAssemblyInformation.IsTestAssembly(assembly) && assembly.assemblyKind == AssemblyInfo.AssemblyKind.Asmdef)
+                if (vai.IsTestAssembly(assembly) && assembly.assemblyKind == AssemblyInfo.AssemblyKind.Asmdef)
                     excludePaths.AddRange(assembly.assembly.sourceFiles);
             }
             string responseFileParameter = string.Empty;
@@ -34,13 +60,6 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests.Standards
                 var excludedPathsParameter = $@"--excluded-paths=""{string.Join(",", excludePaths.Select(s => Path.GetFullPath(s)))}""";
                 File.WriteAllText(responseFilePath, excludedPathsParameter);
                 responseFileParameter = $@"--response-file=""{responseFilePath}""";
-            }
-
-            var filterYamlParameter = "";
-            var filterYamlPath = Path.Combine(packagePath, "Documentation~", "filter.yml");
-            if (Utilities.FileExists(filterYamlPath))
-            {
-                filterYamlParameter = $@"--path-to-filter-yaml=""{filterYamlPath}""";
             }
 
             var startInfo = new ProcessStartInfo(monopath, $@"""{exePath}"" --root-path=""{packagePath}"" {filterYamlParameter} {responseFileParameter}")
@@ -57,42 +76,40 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests.Standards
             process.WaitForExit();
             var stdoutLines = stdout.GetOutput();
             var stderrLines = stderr.GetOutput();
+
+            if (responseFilePath != null)
+                File.Delete(responseFilePath);
+
             if (process.ExitCode != 0)
             {
                 // If FindMissingDocs fails and returns a non-zero exit code (like an unhandled exception) it means that
                 // we couldn't validate the XmdDocValidation because the result is inconclusive. For that reason, we
                 // should add it as an error to be addressed by the developer. If there's any bug with the tool itself
                 // then that will need to be addressed in the XmlDoc repo and rebuild the binaries from PVS.
-                AddError($"FindMissingDocs.exe returned {process.ExitCode}, a non-zero exit code and XmlDocValidation test is inconclusive.");
+                throw new InternalTestErrorException($"FindMissingDocs.exe returned {process.ExitCode}, a non-zero exit code and XmlDocValidation test is inconclusive.");
             }
             if (stderrLines.Length > 0)
             {
-                AddWarning($"Internal Error running FindMissingDocs. Output:\n{string.Join("\n", stderrLines)}");
-                return;
+                throw new InternalTestErrorException($"Internal Error running FindMissingDocs. Output:\n{string.Join("\n", stderrLines)}");
             }
 
-            if (stdoutLines.Length > 0)
-            {
-                var errorMessage = FormatErrorMessage(stdoutLines);
-                AddWarning(errorMessage);
-
-                //// JonH: Enable errors in non-preview packages once the check has been put through its paces and the change is coordinated with RM and PM
-                // if (Context.ProjectPackageInfo.IsPreview)
-                //     Warning(errorMessage);
-                // else
-                // {
-                //     TestState = TestState.Failed;
-                //     Error(errorMessage);
-                // }
-            }
-
-            if (responseFilePath != null)
-                File.Delete(responseFilePath);
+            return stdoutLines;
         }
 
         public static string FormatErrorMessage(IEnumerable<string> expectedMessages)
         {
             return $@"The following APIs are missing documentation: {string.Join(Environment.NewLine, expectedMessages)}";
+        }
+
+        string[] IPvpChecker.Checks => new[] { "PVP-20-1" };
+
+        void IPvpChecker.Run(in PvpRunner.Input input, PvpRunner.Output output)
+        {
+            var filterYamlParameter = ""; // such filtering intentionally not supported in PVP mode
+            foreach (var line in CheckV1Impl(input.Package.path, filterYamlParameter, input.AssemblyInfo, new ValidationAssemblyInformation()))
+            {
+                output.Error("PVP-20-1", line);
+            }
         }
     }
 }